Setup Worker process
sec. 8, video 108 of Docker: The complete guide

First, we create a worker folder and add a package.json.
Add a keys.js to export redisHost & redisPort.

Add an index.js to get a connection over to our Redis server.
Note the retry_strategy is going to tell Redis client that we're using
right here that if it ever loses connection to our Redis server, it
should attempt to automatically reconnect to the server once every 1000
milliseconds.