More Express API setup
sec. 8, video 111 of Docker: The complete guide

Note that we are making a duplicate connection. According to the Redis
documentation for this JavaScript library, if we ever have a client
that's listening or publishing information on Redis, we have to make a
duplicate connection because when a connection is turned into a
connection, that's going to listen or subscribe or publish information,
it can't be used for other purposes. That's why we are doing this
duplicate thing here and other locations (in worker process).

Note that when the route handler receives an index, it will set the hash
of the index ('Nothing yet!') into Redis. Then it will publish a new
'insert' event, which is the one the worker subscribes.
The worker will calculate the Fib number of this index and replace the
'Nothing yet!' string inside of Redis.