Fetch data in the React app
sec. 8, video 114 of Docker: The complete guide

We are going to make 2 separate pages that a user can visit with a
calculator.

In client/src folder, create OtherPage.js.

In client/src folder, create Fib.js
In the componentDidMount function, we fetch the indices and values from
the server and update the state.