Rendering logic in the app
sec. 8, video 115 of Docker: The complete guide

Note that the seenIndexes is an array of objects, the result that we
performed the query to the Postgres. The object contains one prop
number, the field name in the Postgres' table.
The values in the state is an object where the key represents the index
and the value is the Fib number.
We also use controlled input here.

We want handleSubmit to be a bound function, so we use arrow function
syntax.