Add Docker-compose config
sec. 9, video 123 of Docker: The complete guide

Ref: https://hub.docker.com/_/redis

Note that when we specify the dockerfile, we only say Dockerfile.dev
(the name of the dockerfile) instead of specifying any folder like
'./server/Dockerfile.dev'.
To specify that we want to look in the server folder, that's what the
'context' property is for. In the past, we've always made use of context
with dot (.), which meant looking in the current working directory.
e.g. $ docker build -f Dockerfile.dev .
Note that inside of this context is where we are going to also find our
dockerfile as well.

In the volumes, we tell docker that inside the container, don't try to
override the /app/node_modules folder because we don't have node_modules
in our currently directory locally.
Then we copy everything inside of server folder into the app folder of
the container. Anytime we make any change to anything inside that sever
folder will be automatically reflected inside of the app folder of our
container. So we don't have to rebuild our image any time we make a
little change to our source code.