Add environment variables with docker compose
sec. 9, video 124 of Docker: The complete guide

When we setup an environment variable inside of a docker compose file,
we are setting up an environment variable that is applied at runtime.
So only when the container is started up. That information is not going
to being encoded inside the image.
If we just specify the variable's name, then the value for this variable
is going to be taken from our computer.

Remember that any time that we access a service that has been defined
inside of a docker compose file, we just specify the name of the
service.
Notice that the default port of Redis server is 6379 and the default
port for Postgres is 5432.

We can use the following command to start containers
$ docker-compose up

Note that in our case, the Postgres starts very slowly so the server
failed to connect to it. We can use restart command to overcome this
issue. But a better solution is to configure the Postgres client to
automatically attempt reconnections to database.

https://www.udemy.com/course/docker-and-kubernetes-the-complete-guide/learn/lecture/11437290#questions/5102100