Routing with Nginx
sec. 9, video 126 & 127 of Docker: The complete guide

Notice that we made absolutely no mention of any ports inside this
docker-compose.yml. So nothing is exposed to the outside world.

Note that in the client, we send request to '/api/values/…'. However,
in the server, our api starts with '/values/…'.

The purpose of Nginx server here is that anytime a request comes into
it, it's going to look at the incoming request's 'request path'.
If '/api/' is in the route that this request is trying to access, then
it will redirect this request over to the Express server.
Otherwise, it will direct that request over the React server instead.
So we can just expose the Nginx service.

The reason we don't assign ports to React server and express server is
that the port can be easily changed.
After Nginx does routing step, it will chop off the '/api' part of the
routes. It means we can specify these routes with configuration that
only really changes at the Nginx layer. We don't have to customize the
routes defined in the Express server based upon whether it's supposed to
be watching for a '/api' or not.

Note that in the server, we are listening on port 5000.

And the client has it create-react-app application by default listens
for traffic on port 3000.
The client in 'client:3000' and server in 'server:5000' are both actual
addresses (host names or URL) that Nginx is going to try to direct
traffic to. Those the names of the client and server services we put
together inside of our docker-compose file.

Create nginx folder and add 'default.conf' into it. Note the name should
be 'default.conf'. It is a special file name.
Because 'server' is a keyword of Nginx, we are going to change our
terminology for our back-end service to 'api'.

We apply a regex of /api against the path name. The $1 is a reference to
whatever text was matched by (.*).
The break keyword is a directive and it essentially means that do not
try to apply any other rewrite rules after applying this one.