Build a custom Nginx image
sec. 9, video 130 of Docker: The complete guide

Ref: https://hub.docker.com/_/nginx

In order to do any customization of Nginx, essentially all we have to do
is build a Dockerfile that says we want to make an image from Nginx and
then we're going to take our configuration file and copy it into the
image. As long as we copy that configuration file into our image, it
will be automatically applied to Nginx once a container starts up out of
our custom image.

Inside of the nginx folder, create a Dockerfile.dev file. We're saying
very clearly that take our file and overwrite the existing one.

Add nginx service into docker-compose.yml.