Fix for 'I Calculated Nothing Yet' message
sec. 9, video 132 of Docker: The complete guide

It's entirely possible that you might run into a bug or two when you
start up this set of containers with docker-compose.
Are you able to enter a number into the react app, but it appears to
never be calculated, as seen in this screenshot?
If that's the case, then you can try adding in environment variables to
the 'worker' entry in the docker-compose file.
Also, you can add depends_on to the api.

https://docs.docker.com/compose/compose-file/#depends_on
In our case, this still does not work. So we add restart: always to api
service to fix it.

After any changes, make sure you run
$ docker-compose down
and then
$ docker-compose up --build