Open WebSocket connections
sec. 9, video 133 & 134 of Docker: The complete guide

If we open the dev view on Chrome, we can find a WebSocket error.
Essentially we're seeing this error message because any time our
React-app boots up in development mode, it wants to keep an active
connection to the development server and be notified any time that some
file changes.

We are going to expose one route through the Nginx server that will
allow a websocket connection to be made with the running React process.
Note that this is really a development specific piece of configuration.