Add production Dockerfiles
sec. 10, video 136 of Docker: The complete guide

Create a Dockerfile in worker folder. Copy everything in Dockerfile.dev
and paste it. Change the default command to "npm run start".
Note that we have a corresponding 'start script' in the package.json.
Then do the same thing for server.

Do the same thing for Nginx.
Note that for Nginx, we have a development specific piece of
configuration for WebSocket. If we really want to have a separate
production version, we can make a separate default.conf that would not
have the WebSocket block inside of it.