Alter Nginx's listen port
sec. 10, video 137 & 138 of Docker: The complete guide

When we build our production Dockerfile for the React app, we need to
customize Nginx to make sure that listens on port 3000.

Create a nginx folder in the client folder. Create a default.conf. Note
that we will setup an Nginx server that hosts the client project.
First, we make Nginx listen on port 3000 and serves up the React
production from there instead of the default of port 80.
Then we also setup a routing rule. Anytime someone comes to the root
route, we are going to set a root /usr/share/nginx/html, that's where
we're going to put all of our production asets.
We also setup an index directive. That's going to be our index.html or
index.htm files.
Create a Dockerfile for production in client folder. This is a
multi-stage build process.