Clean up tests
sec. 10, video 140 of Docker: The complete guide

Currently, we have a default test that tries to assert that we can
render out the <App> component without anything crashing during the
process.
But if we run this test now, we would probably see crash. The reason is
that the <App> component now attempts to render out the <Fib> component.
And the <Fib> component is going to try to make a request to our
back-end Express server. That's not quite running at this point in time.
To make sure that our tests run successfully, we are going to delete the
3 lines of test in there.
If we were running a real test suite right here rather than making a
request off to the real Express API, what we'd probably do is set up a
little faker module to kind of mark out that request and return some
dummy JSON very quickly.