Setup Travis configuration
Create .travis.yml in the root folder.
As a before_install step, we are going to build a test version of our
client project and then eventually run some tests inside of it.
To build out a test version of the client project, we'll make an image
out of the Dockerfile.dev. We're making use of the development
Dockerfile because we only get access to all there test suites and all
that kind of stuff. When we have all of the dependencies attached the
project.
The production Dockerfile installs dependencies, builds a production
version of the project. And it copies over just the very raw production
version of our assets. So that's why we're not going to use the
production Dockerfile. It does not allow us to run any test.
Note that we need to specify the build context ./client.

The script section is the primary test running section. If any of the
scripts that we add inside of here exits with the status code other than
zero, then Travis CI is going to assume that our build failed.
If we have other project that we want to run tests, we can add in and
additional build and install script in before_install section. And run
the test in script section. We can add in as many scripts as we want to.

In the after_success section, we will add the configuration that's going
to build a production version of all of our projects and then push then
off to Docker Hub.