Push images to Docker Hub
sec. 10, video 143 of Docker: The complete guide

Before we attempt to push those images off to Docker Hub, we need to
first log in to the Docker CLI. We're going to setup some encrypted
environment variables that are only accessible by Travis CI. And we'll
add in our username and password over there.

In .travis.yml, we need to log in to the docker CLI.
First we retrieve our docker password from the environment variable
(echo "DOCKER_PASSWORD") and essentially emit that over standard in as
input to the next command (the command on the other side of the |).
We then run the "docker login" command and add our user name as a -u
flag. And then we tell "docker login " that it can expect to receive
that password over standard in.
Essentially this is how we log into Docker in one single command.

Then we commit the change to .travis.yml and push the commit to Git Hub.